+++ homepage/homepage/app/(others)/examples/page.tsx
@@ -6,11 +6,11 @@ import { ReactNativeLogo } from "@/components/icons/ReactNativeLogo";
 import { SvelteLogo } from "@/components/icons/SvelteLogo";
 import { VueLogo } from "@/components/icons/VueLogo";
 import { Example, features, tech } from "@/content/example";
-import { clsx } from "clsx";
 import { H2 } from "@garden-co/design-system/src/components/atoms/Headings";
 import { Icon } from "@garden-co/design-system/src/components/atoms/Icon";
 import { GappedGrid } from "@garden-co/design-system/src/components/molecules/GappedGrid";
 import { HeroHeader } from "@garden-co/design-system/src/components/molecules/HeroHeader";
+import { clsx } from "clsx";
 import type { Metadata } from "next";
 
 const title = "Examples";
@@ -198,6 +198,12 @@ const MusicIllustration = () => (
   </div>
 );
 
+const JazzPaperScissorsIllustration = () => (
+  <div className="flex flex-col items-center justify-center h-full p-8 text-4xl">
+    ✊ ✋ ✌️
+  </div>
+);
+
 const ImageUploadIllustration = () => (
   <div className="flex flex-col items-center justify-center h-full p-8">
     <div className="p-3 w-[12rem] h-[8rem] border border-dashed border-blue dark:border-blue-500 rounded-lg flex gap-2 flex-col items-center justify-center">
@@ -255,21 +261,21 @@ const ReactionsIllustration = () => (
 
 const MultiCursorIllustration = () => (
   <div className="flex bg-stone-100 h-full flex-col items-center justify-center dark:bg-transparent p-4">
-    <div className=" bg-white md:aspect-[3/2] flex flex-col rounded-md shadow-xl shadow-stone-400/20 dark:shadow-none">
+    <div className=" bg-white min-w-64 md:aspect-[3/2] flex flex-col rounded-md shadow-xl shadow-stone-400/20 dark:shadow-none">
       <div className="w-full py-2 flex items-center gap-1.5 px-2 border-b dark:border-b-stone-200">
         <span className="rounded-full size-2 bg-stone-200"></span>
         <span className="rounded-full size-2 bg-stone-200"></span>
         <span className="rounded-full size-2 bg-stone-200"></span>
       </div>
 
-      <div className="h-full mx-auto flex flex-col justify-center p-12 sm:p-16">
+      <div className="h-full mx-auto flex flex-col justify-center p-12">
         <div className="inline-block relative px-1 ring-1 ring-blue-400">
           <div className="absolute size-2 bg-white border border-blue-400 -left-1 -top-1"></div>
           <div className="absolute size-2 bg-white border border-blue-400 -right-1 -top-1"></div>
           <div className="absolute size-2 bg-white border border-blue-400 -left-1 -bottom-1"></div>
           <div className="absolute size-2 bg-white border border-blue-400 -right-1 -bottom-1"></div>
 
-          <span className="text-lg font-semibold md:text-2xl md:font-bold text-stone-800 ">
+          <span className="text-lg font-semibold md:text-2xl md:font-bold text-stone-800">
             Hello, world!
           </span>
           <div className="absolute -top-10 right-4 text-rose-600 flex items-end gap-1">
@@ -280,6 +286,21 @@ const MultiCursorIllustration = () => (
             <span className="text-xs">Sebastian</span>
           </div>
         </div>
+      </div>
+    </div>
+  </div>
+);
+
+const CoTextIllustration = () => (
+  <div className="flex bg-stone-100 h-full flex-col items-center justify-center dark:bg-transparent p-4">
+    <div className=" bg-white md:aspect-[3/2] min-w-64 flex flex-col rounded-md shadow-xl shadow-stone-400/20 dark:shadow-none">
+      <div className="flex gap-2 p-3 border-b">
+        <Icon name="bold" size="xs" />
+        <Icon name="italic" size="xs" />
+        <Icon name="code" size="xs" />
+      </div>
+      <div className="py-2 px-3 text-xl text-stone-800">
+        <em>Hello</em>, <strong>world!</strong>
       </div>
     </div>
   </div>
@@ -434,8 +455,18 @@ const reactExamples: Example[] = [
       "Track user presence on a canvas with multiple cursors and out of bounds indicators.",
     tech: [tech.react],
     features: [features.coFeed],
-    demoUrl: "https://jazz-multi-cursors.vercel.app",
+    demoUrl: "https://multi-cursors-demo.jazz.tools",
     illustration: <MultiCursorIllustration />,
+  },
+  {
+    name: "Collaborative rich text",
+    slug: "richtext",
+    description:
+      "Handle multiple users editing the same text, integrated with a ProseMirror editor for rich text.",
+    tech: [tech.react],
+    features: [features.coRichText, features.coPlainText],
+    demoUrl: "https://richtext-demo.jazz.tools",
+    illustration: <CoTextIllustration />,
   },
   {
     name: "Rate my pet",
@@ -478,6 +509,16 @@ const reactExamples: Example[] = [
     illustration: <MusicIllustration />,
   },
   {
+    name: "Jazz paper scissors",
+    slug: "jazz-paper-scissors",
+    description:
+      "A game that shows how to communicate with other accounts through the experimental Inbox API.",
+    tech: [tech.react],
+    features: [features.serverWorker, features.inbox],
+    illustration: <JazzPaperScissorsIllustration />,
+    demoUrl: "https://jazz-paper-scissors.vercel.app"
+  },
+  {
     name: "Clerk",
     slug: "clerk",
     description: "A React app that uses Clerk for authentication",
@@ -511,6 +552,7 @@ const reactExamples: Example[] = [
     tech: [tech.react],
     features: [features.inviteLink],
     illustration: <OrganizationIllustration />,
+    demoUrl: "https://jazz-organization.vercel.app"
   },
   {
     name: "Version history",
@@ -519,6 +561,7 @@ const reactExamples: Example[] = [
       "Track and restore previous versions of your data, and see who made the changes.",
     tech: [tech.react],
     illustration: <VersionHistoryIllustration />,
+    demoUrl: "https://jazz-version-history.vercel.app"
   },
 ];
 
@@ -566,7 +609,6 @@ const vueExamples: Example[] = [
     description: "A todo list where you can collaborate with invited guests.",
     tech: [tech.vue],
     features: [features.inviteLink],
-    demoUrl: "https://todo-demo.jazz.tools",
     illustration: (
       <div className="h-full w-full bg-cover bg-[url('/todo.jpg')] bg-left-bottom"></div>
     ),