+++ homepage/homepage/app/layout.tsx
@@ -1,11 +1,11 @@
 import "./globals.css";
-import type { Metadata } from "next";
-import { fontClasses } from "@garden-co/design-system/src/fonts";
 import { ThemeProvider } from "@/components/ThemeProvider";
 import { JazzFooter } from "@/components/footer";
 import { marketingCopy } from "@/content/marketingCopy";
+import { fontClasses } from "@garden-co/design-system/src/fonts";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
+import type { Metadata } from "next";
 
 const metaTags = {
   title: `Jazz - ${marketingCopy.headline}`,