+++ examples/chat-rn/CHANGELOG.md
@@ -1,5 +1,15 @@
 # chat-rn
 
+## 1.0.103
+
+### Patch Changes
+
+- Updated dependencies [c6fb8dc]
+  - cojson@0.13.16
+  - cojson-transport-ws@0.13.16
+  - jazz-react-native@0.13.16
+  - jazz-tools@0.13.16
+
 ## 1.0.102
 
 ### Patch Changes