+++ examples/passkey/CHANGELOG.md
@@ -1,5 +1,12 @@
 # minimal-auth-passkey
 
+## 0.0.82
+
+### Patch Changes
+
+- jazz-react@0.13.16
+- jazz-tools@0.13.16
+
 ## 0.0.81
 
 ### Patch Changes