+++ packages/cojson-storage-indexeddb/CHANGELOG.md
@@ -1,5 +1,13 @@
 # cojson-storage-indexeddb
 
+## 0.13.16
+
+### Patch Changes
+
+- Updated dependencies [c6fb8dc]
+  - cojson@0.13.16
+  - cojson-storage@0.13.16
+
 ## 0.13.15
 
 ### Patch Changes