+++ packages/jazz-inspector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # jazz-inspector
 
+## 0.13.16
+
+### Patch Changes
+
+- Updated dependencies [c6fb8dc]
+  - cojson@0.13.16
+  - jazz-react-core@0.13.16
+  - jazz-tools@0.13.16
+
 ## 0.13.15
 
 ### Patch Changes