+++ packages/jazz-react-native-media-images/CHANGELOG.md
@@ -1,5 +1,11 @@
 # jazz-browser-media-images
 
+## 0.13.16
+
+### Patch Changes
+
+- jazz-tools@0.13.16
+
 ## 0.13.15
 
 ### Patch Changes