+++ packages/jazz-richtext-prosemirror/CHANGELOG.md
@@ -1,5 +1,12 @@
 # jazz-richtext-prosemirror
 
+## 0.1.5
+
+### Patch Changes
+
+- jazz-browser@0.13.16
+- jazz-tools@0.13.16
+
 ## 0.1.4
 
 ### Patch Changes